Separate TreeMenu out into its own class.
