+++ pytripgui/controller/tree_cont.py
@@ -8,8 +8,10 @@ from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QMenu
 from PyQt5.QtWidgets import QTreeWidgetItem
 
-import pytrip as pt
-import pytrip.tripexecuter as pte
+# import pytrip as pt
+# import pytrip.tripexecuter as pte
+
+from pytripgui.controller.tree_menu_cont import TreeMenuController
 
 logger = logging.getLogger(__name__)
 
@@ -27,7 +29,7 @@ class TreeController(object):
         """
         self.model = model
         self.view = view
-        tw = view.treeWidget
+        # tw = view.treeWidget
 
         # QTreeWidgetItem placeholders
         self.tctx = None
@@ -37,33 +39,12 @@ class TreeController(object):
         self.tdos = []
         self.tlet = []
 
-        tw.setContextMenuPolicy(Qt.CustomContextMenu)
-        tw.customContextMenuRequested.connect(self.on_context_menu)
+        # setup the submenus:
+        self.tmc = TreeMenuController(model, view)  # self, else it goes out of scope?
 
     def on_checked_changed(self, pos):
         logger.debug("on_checked_changed() {}".format(pos))
         # tw = self.view.tw
-
-    def on_context_menu(self, pos):
-        """
-        Callback if rightclick on treeWidget.
-        """
-
-        logger.debug("on_context_menu() {}".format(pos))
-
-        tw = self.view.treeWidget
-
-        getSelected = tw.selectedItems()
-        if getSelected:
-            baseNode = getSelected[0]  # QTreeWidgetItem
-            obj = baseNode.data(0, Qt.UserRole)
-        else:
-            obj = pt.CtxCube  # if nothing selected, put up Ctx menu
-
-        popup_menu = self.create_popup_menu(obj)
-        node = tw.mapToGlobal(pos)
-
-        action = popup_menu.exec_(node)
 
     def update_tree(self):
         """
@@ -145,62 +126,3 @@ class TreeController(object):
                     return child2
 
         return None
-
-    def create_popup_menu(self, obj):
-        """
-        obj is the data object for the QTreeWidgetItem (CtxCube VdxCube ... etc)
-        """
-        tw = self.view.treeWidget
-        popup_menu = QMenu(tw)
-
-        if isinstance(obj, pt.CtxCube):
-            popup_menu.addAction("Open DICOM", self.menu_open)
-            popup_menu.addAction("Open .ctx", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addAction("Export .ctx", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        if isinstance(obj, pt.VdxCube):
-            popup_menu.addAction("New", self.menu_open)
-            popup_menu.addAction("Open .vdx", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addAction("Export .vdx", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete all", self.menu_open)
-
-        if isinstance(obj, pt.Voi):
-            popup_menu.addAction("Color", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        if isinstance(obj, pte.Plan):
-            popup_menu.addAction("Edit", self.menu_open)
-            popup_menu.addAction("Export", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        if isinstance(obj, pte.Field):
-            popup_menu.addAction("Edit", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        if isinstance(obj, pt.DosCube):
-            popup_menu.addAction("Export", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        if isinstance(obj, pt.LETCube):
-            popup_menu.addAction("Export", self.menu_open)
-            popup_menu.addAction("Rename", self.menu_open)
-            popup_menu.addSeparator()
-            popup_menu.addAction("Delete", self.menu_open)
-
-        return popup_menu
-
-    def menu_open(self, event):
-        logger.debug("menu_open() {}".format(event))
